Working Code before Base Pipeline changes
