+++ tile_engine/CMakeLists.txt
@@ -5,4 +5,4 @@ include_directories(BEFORE
         ${CMAKE_CURRENT_LIST_DIR}/include
     )
 
-add_subdirectory(ops)
+add_subdirectory(ops/gemm)
