--- tile_engine/ops/CMakeLists.txt
@@ -1,7 +1,0 @@
-# Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
-# SPDX-License-Identifier: MIT
-
-add_subdirectory(gemm)
-add_subdirectory(gemm_multi_d)
-add_subdirectory(gemm_preshuffle)
-add_subdirectory(gemm_streamk)
