+++ tile_engine/ops/gemm/gemm_instance_builder.py
@@ -1,17 +1,9 @@
-#!/usr/bin/env python
-# Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
-# SPDX-License-Identifier: MIT
-
-
 import os
 import json
-import argparse
-import itertools
-import multiprocessing
-import concurrent.futures
 from pathlib import Path
-import logging
 import importlib.util
+import itertools
+import logging
 
 
 def _import_validation_utils():
@@ -22,7 +14,7 @@ def _import_validation_utils():
     # Load the module dynamically
     spec = importlib.util.spec_from_file_location(
         "validation_utils",
-        os.path.join(parent_dir, "commons", "gemm_validation_utils.py"),
+        os.path.join(parent_dir, "gemm", "gemm_validation_utils.py"),
     )
     validation_utils = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(validation_utils)
@@ -34,14 +26,22 @@ def _import_validation_utils():
 _validation_utils = _import_validation_utils()
 is_tile_config_valid = _validation_utils.is_tile_config_valid
 is_trait_combination_valid = _validation_utils.is_trait_combination_valid
-get_dtype_string = _validation_utils.get_dtype_string
 get_abc_layouts = _validation_utils.get_abc_layouts
-
-logging.basicConfig(level=logging.INFO)
+get_abcd_layouts = _validation_utils.get_abcd_layouts
+get_dtype_string = _validation_utils.get_dtype_string
 
 
 class GemmKernelBuilder:
-    def __init__(self, working_path, gpu_target, datatype, layout, config_json=None):
+    def __init__(
+        self,
+        kernel_name_prefix,
+        working_path,
+        gpu_target,
+        datatype,
+        layout,
+        config_json=None,
+    ):
+        self.kernel_name_prefix = kernel_name_prefix
         self.working_path = Path(working_path)
         self.gpu_target = gpu_target
         self.datatype = datatype
@@ -56,10 +56,10 @@ class GemmKernelBuilder:
             with open(config_json, "r") as f:
                 self.config = json.load(f)
 
-    def write_kernel_list(self):
+    def _list_kernels(self):
         """Write kernel list to file for CMake to read (with comprehensive validation)"""
         # Get configurations using comprehensive validation
-        tile_configs = self._get_tile_configs(fast_mode=False)
+        tile_configs = self._get_tile_configs(self.kernel_name_prefix)
         trait_combos = self._generate_trait_combinations()
 
         kernel_list = []
@@ -76,7 +76,7 @@ class GemmKernelBuilder:
                 ) = trait_combo
 
                 # Create kernel name with proper boolean capitalization
-                kernel_name = f"gemm_{self.datatype}_{self.layout}_{pipeline}_{epilogue}_{scheduler}_{str(pad_m).capitalize()}_{str(pad_n).capitalize()}_{str(pad_k).capitalize()}_{str(persistent).capitalize()}"
+                kernel_name = f"{self.kernel_name_prefix}_{self.datatype}_{self.layout}_{pipeline}_{epilogue}_{scheduler}_{str(pad_m).capitalize()}_{str(pad_n).capitalize()}_{str(pad_k).capitalize()}_{str(persistent).capitalize()}"
 
                 # Create tile configuration string
                 tile_str = f"{tile_config['tile_m']}x{tile_config['tile_n']}x{tile_config['tile_k']}_"
@@ -94,11 +94,15 @@ class GemmKernelBuilder:
                 )
 
         # Write kernel count
-        with open(self.working_path / "gemm_kernel_count.txt", "w") as f:
+        with open(
+            self.working_path / f"{self.kernel_name_prefix}_kernel_count.txt", "w"
+        ) as f:
             f.write(str(len(kernel_list)))
 
         # Write kernel list
-        with open(self.working_path / "gemm_kernel_list.txt", "w") as f:
+        with open(
+            self.working_path / f"{self.kernel_name_prefix}_kernel_list.txt", "w"
+        ) as f:
             for kernel in kernel_list:
                 # Format: kernel_name|tile_config|trait_combo
                 tile_config = kernel["tile_config"]
@@ -119,6 +123,7 @@ class GemmKernelBuilder:
 
     def _get_tile_configs(self, fast_mode=False):
         """Get tile configurations for the current datatype and layout"""
+
         tile_config = self.config["tile_config"]
 
         # Generate values in the config if default range is given
@@ -153,6 +158,14 @@ class GemmKernelBuilder:
         warp_tile_k_values = tile_config.get("warp_tile_k").get("values")
 
         # Generate all combinations
+        default_pipeline = ""
+        if self.kernel_name_prefix == "gemm_universal":
+            default_pipeline = "compv4"
+        elif self.kernel_name_prefix == "gemm_multi_d":
+            default_pipeline = "compv4"
+        elif self.kernel_name_prefix == "gemm_preshuffle":
+            default_pipeline = "preshufflev2"
+
         configs = []
         for tile_m in tile_m_values:
             for tile_n in tile_n_values:
@@ -174,6 +187,7 @@ class GemmKernelBuilder:
                                                 warp_tile_m,
                                                 warp_tile_n,
                                                 warp_tile_k,
+                                                default_pipeline,
                                                 fast_mode=fast_mode,
                                             ):
                                                 configs.append(
@@ -211,8 +225,8 @@ class GemmKernelBuilder:
         warp_tile_m,
         warp_tile_n,
         warp_tile_k,
-        pipeline="compv4",  # Default pipeline for validation
-        fast_mode=False,  # Add fast mode option
+        pipeline,
+        fast_mode=False,
     ):
         """Validate that tile configuration is reasonable"""
         if fast_mode:
@@ -234,6 +248,15 @@ class GemmKernelBuilder:
 
             return True
         else:
+            # Validate preshuffle specific constraints
+            if (
+                self.config.get("permute_n") is not None
+                and self.config.get("permute_n") is True
+            ):
+                valid = (tile_n / warp_tile_n / warp_n) % 2 == 0
+                if not valid:
+                    return False
+
             # Full validation for generation
             # Determine data types for validation
             a_datatype = self.datatype
@@ -302,10 +325,13 @@ class GemmKernelBuilder:
                 )
         return combinations
 
-    def _generate_kernel_instance(
-        self, tile_config, trait_combo, k_block_per_cu, is_header=True
-    ):
+    def _generate_kernel_instance(self, tile_config, trait_combo):
         """Generate a single kernel instance"""
+
+        k_block_per_cu = self.config.get("k_block_per_cu")
+        if k_block_per_cu is None:
+            k_block_per_cu = 1
+
         (
             pipeline,
             epilogue,
@@ -317,7 +343,7 @@ class GemmKernelBuilder:
         ) = trait_combo
 
         # Create kernel name with proper boolean capitalization
-        kernel_name = f"gemm_{self.datatype}_{self.layout}_{pipeline}_{epilogue}_{scheduler}_{str(pad_m).capitalize()}_{str(pad_n).capitalize()}_{str(pad_k).capitalize()}_{str(persistent).capitalize()}"
+        kernel_name = f"{self.kernel_name_prefix}_{self.datatype}_{self.layout}_{pipeline}_{epilogue}_{scheduler}_{str(pad_m).capitalize()}_{str(pad_n).capitalize()}_{str(pad_k).capitalize()}_{str(persistent).capitalize()}"
 
         # Create tile configuration string
         tile_str = (
@@ -330,35 +356,76 @@ class GemmKernelBuilder:
 
         kernel_name += f"_{tile_str}"
 
-        # Map pipeline names to the correct pipeline implementation
-        pipeline_impl_map = {
-            "mem": "ck_tile::GemmPipelineAgBgCrMem",
-            "compv3": "ck_tile::GemmPipelineAgBgCrCompV3",
-            "compv4": "ck_tile::GemmPipelineAgBgCrCompV4",
-        }
+        if self.kernel_name_prefix in ["gemm_universal", "gemm_multi_d"]:
+            # Map pipeline names to the correct pipeline implementation
+            pipeline_impl_map = {
+                "mem": "ck_tile::GemmPipelineAgBgCrMem",
+                "compv3": "ck_tile::GemmPipelineAgBgCrCompV3",
+                "compv4": "ck_tile::GemmPipelineAgBgCrCompV4",
+            }
+            # Map pipeline names to base pipeline for hot loop detection
+            base_pipeline_map = {
+                "mem": "ck_tile::BaseGemmPipelineAgBgCrMem",
+                "compv3": "ck_tile::BaseGemmPipelineAgBgCrCompV3",
+                "compv4": "ck_tile::BaseGemmPipelineAgBgCrCompV4",
+            }
+        elif self.kernel_name_prefix == "gemm_preshuffle":
+            # Map pipeline names to the correct pipeline implementation
+            pipeline_impl_map = {
+                "preshufflev2": "ck_tile::WeightPreshufflePipelineAGmemBGmemCRegV2",
+            }
+            # Map pipeline names to base pipeline for hot loop detection
+            base_pipeline_map = {
+                "preshufflev2": "ck_tile::BaseWeightPreshufflePipelineAGmemBGmemCRegV2",
+            }
 
-        # Map scheduler names to the correct enum values
         scheduler_type_map = {
             "intrawave": "ck_tile::GemmPipelineScheduler::Intrawave",
             "interwave": "ck_tile::GemmPipelineScheduler::Interwave",
             "default": "ck_tile::GemmPipelineScheduler::Default",
         }
 
-        # Determine accumulator type based on datatype
-        acc_type = "float"
+        instance_code = self.populate_kernel_header(kernel_name)
+        instance_code += self.populate_kernel_dtype_layout(self.kernel_name_prefix)
+        instance_code += self.populate_strut_begin(kernel_name)
+        instance_code += self.populate_tile_config(tile_config)
+        instance_code += self.populate_trait_config(
+            self.kernel_name_prefix, trait_combo
+        )
+        instance_code += self.populate_initialization(
+            self.kernel_name_prefix, base_pipeline_map, pipeline
+        )
+        instance_code += self.populate_launch(
+            self.kernel_name_prefix,
+            scheduler_type_map,
+            scheduler,
+            pipeline_impl_map,
+            pipeline,
+            epilogue,
+            k_block_per_cu,
+            persistent,
+        )
 
-        # Determine output type
-        c_type = self.datatype
-        if self.datatype in ["fp8", "bf8"]:
-            c_type = "fp16"
+        # Write into a file
+        simplified_name = kernel_name
+        if simplified_name.startswith(f"{self.kernel_name_prefix}_"):
+            simplified_name = simplified_name[len(self.kernel_name_prefix) + 1 :]
 
-        # Determine layouts based on self.layout
-        a_layout, b_layout, c_layout = get_abc_layouts(self.layout)
+        header_file = (
+            self.working_path
+            / f"{self.kernel_name_prefix}_single_{simplified_name}.hpp"
+        )
+        with open(header_file, "w") as f:
+            f.write(instance_code)
 
-        # Generate kernel instance code using the correct API
-        pragma_line = "#pragma once\n" if is_header else ""
+        print(f"Generated {header_file}")
+
+        return kernel_name, instance_code
+
+    def populate_kernel_header(self, kernel_name):
         instance_code = f"""// Generated kernel instance for {kernel_name}
-{pragma_line}
+#pragma once
+
 #include <cstdint>
 #include <utility>
 #include <tuple>
@@ -369,22 +436,66 @@ class GemmKernelBuilder:
 #include "ck_tile/ops/common/tensor_layout.hpp"
 #include "ck_tile/ops/epilogue/default_2d_epilogue.hpp"
 #include "ck_tile/ops/epilogue/cshuffle_epilogue.hpp"
+"""
+        return instance_code
 
+    def populate_kernel_dtype_layout(self):
+        # Determine accumulator type based on datatype
+        acc_type = "float"
+
+        # Determine output type
+        c_type = self.datatype
+        if self.datatype in ["fp8", "bf8"]:
+            c_type = "fp16"
+
+        # Determine layouts based on self.layout
+        if self.kernel_name_prefix == "gemm_multi_d":
+            a_layout, b_layout, c_layout, ds_layout = get_abcd_layouts(self.layout)
+        elif (
+            self.kernel_name_prefix == "gemm_universal"
+            or self.kernel_name_prefix == "gemm_preshuffle"
+        ):
+            a_layout, b_layout, c_layout = get_abc_layouts(self.layout)
+
+        instance_code = f"""
 using ADataType = {get_dtype_string(self.datatype)};
 using BDataType = {get_dtype_string(self.datatype)};
 using AccDataType = {acc_type};
-using CDataType = {get_dtype_string(c_type)};
+using CDataType = {get_dtype_string(c_type)};"""
 
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code += f"""
+using D0DataType = {get_dtype_string(self.datatype)};
+using D1DataType = {get_dtype_string(self.datatype)};
+using DsDataType = ck_tile::tuple<D0DataType, D1DataType>;"""
+
+        instance_code += f"""
 using ALayout = {a_layout};
 using BLayout = {b_layout};
 using CLayout = {c_layout};
+"""
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code += f"""
+using D0Layout = {ds_layout[0]};
+using D1Layout = {ds_layout[1]};
+using DsLayout = ck_tile::tuple<D0Layout, D1Layout>;
 
+using ElementWiseFn = ck_tile::element_wise::{self.elementwise_function};"""
+
+        return instance_code
+
+    def populate_strut_begin(self, kernel_name):
+        instance_code = f"""
 // Kernel name for display
 constexpr const char* KERNEL_NAME = "{kernel_name}";
 
 // Wrapper for simplified launch interface
 struct SelectedKernel {{
-    // Tile configuration
+    """
+        return instance_code
+
+    def populate_tile_config(self, tile_config):
+        instance_code = f"""// Tile configuration
     static constexpr ck_tile::index_t BlockSize = 256;
     static constexpr ck_tile::index_t TileM = {tile_config["tile_m"]};
     static constexpr ck_tile::index_t TileN = {tile_config["tile_n"]};
@@ -394,34 +505,163 @@ struct SelectedKernel {{
     static constexpr ck_tile::index_t WarpPerBlock_K = {tile_config["warp_k"]};
     static constexpr ck_tile::index_t WarpTileM = {tile_config["warp_tile_m"]};
     static constexpr ck_tile::index_t WarpTileN = {tile_config["warp_tile_n"]};
-    static constexpr ck_tile::index_t WarpTileK = {tile_config["warp_tile_k"]};
+    static constexpr ck_tile::index_t WarpTileK = {tile_config["warp_tile_k"]};"""
+        return instance_code
 
-    // Traits
+    def populate_trait_config(self, trait_combo):
+        (
+            pipeline,
+            epilogue,
+            scheduler,
+            pad_m,
+            pad_n,
+            pad_k,
+            persistent,
+        ) = trait_combo
+
+        instance_code = f"""
+    
+    // Traits configurations
     static constexpr bool kPadM = {"true" if pad_m in [True, "true"] else "false"};
     static constexpr bool kPadN = {"true" if pad_n in [True, "true"] else "false"};
     static constexpr bool kPadK = {"true" if pad_k in [True, "true"] else "false"};
     static constexpr bool TransposeC = false;
+    static constexpr bool DoubleSmemBuffer = {"true" if pipeline in ["compv4", "preshufflev2"] else "false"};"""
+
+        if self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code += f"""    
     static constexpr bool UsePersistentKernel = {"true" if persistent in [True, "true"] else "false"};
-    static constexpr bool DoubleSmemBuffer = {"true" if pipeline == "compv4" else "false"};
     static constexpr bool UseStructuredSparsity = false;
-    static constexpr bool Preshuffle = false;
-    static constexpr ck_tile::index_t NumWaveGroups = 1;
+    static constexpr ck_tile::index_t NumWaveGroups = 1;"""
+
+            if self.kernel_name_prefix == "gemm_preshuffle":
+                instance_code += f"""
+    static constexpr bool Preshuffle = true;
+    static constexpr bool PermuteN     = {"true" if self.config.get("permute_n") else "false"};"""
+            else:
+                instance_code += """
+    static constexpr bool Preshuffle = false;"""
+        return instance_code
+
+    def populate_initialization(self, base_pipeline_map, pipeline):
+        # Tile Shape
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code = """  
 
     // Tile shape
     using TileShape = ck_tile::TileGemmShape<
         ck_tile::sequence<TileM, TileN, TileK>,
         ck_tile::sequence<WarpPerBlock_M, WarpPerBlock_N, WarpPerBlock_K>,
-        ck_tile::sequence<WarpTileM, WarpTileN, WarpTileK>,
-        false, false>;
-    
-    // Tile partitioner
-    using TilePartitioner = ck_tile::GemmSpatiallyLocalTilePartitioner<TileShape, 8, 4>;
+        ck_tile::sequence<WarpTileM, WarpTileN, WarpTileK>>;"""
 
-    static float launch(const ck_tile::GemmHostArgs& args, const ck_tile::stream_config& stream) {{
+        elif self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code = """
+            
+    // Tile shape
+    using TileShape = ck_tile::TileGemmShape<
+        ck_tile::sequence<TileM, TileN, TileK>,
+        ck_tile::sequence<WarpPerBlock_M, WarpPerBlock_N, WarpPerBlock_K>,
+        ck_tile::sequence<WarpTileM, WarpTileN, WarpTileK>,
+        false, false>;"""
+
+        # Tile partitioner
+        instance_code += """
+
+    // Tile partitioner
+    using TilePartitioner = ck_tile::GemmSpatiallyLocalTilePartitioner<TileShape, 8, 4>;"""
+
+        # Traits
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code += """
+
+    // Traits
+    using Traits = ck_tile::TileGemmTraits<kPadM, kPadN, kPadK, ALayout, BLayout, CLayout>;"""
+        elif self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code += """
+
+    // Traits
+    using Traits = ck_tile::TileGemmTraits<kPadM, kPadN, kPadK, ALayout, BLayout, CLayout, NumWaveGroups>;"""
+
+        # Pipeline problem
+        instance_code += """
+
+    // Pipeline problem
+    using GemmPipelineProblem = ck_tile::GemmPipelineProblem<
+        ADataType,
+        BDataType,
+        AccDataType,
+        TileShape,
+        Traits>;"""
+
+        # Base pipeline for hot loop detection
+        if self.kernel_name_prefix == "gemm_preshuffle":
+            instance_code += f"""
+
+    // Base pipeline for hot loop detection
+    using BaseGemmPipeline = {base_pipeline_map.get(pipeline, "ck_tile::BaseWeightPreshufflePipelineAGmemBGmemCRegV2")}<GemmPipelineProblem>;"""
+
+        elif self.kernel_name_prefix in ["gemm_universal", "gemm_multi_d"]:
+            instance_code += f"""
+
+    // Base pipeline for hot loop detection
+    using BaseGemmPipeline = {base_pipeline_map.get(pipeline)}<GemmPipelineProblem>;"""
+
+        return instance_code
+
+    def populate_launch(
+        self,
+        scheduler_type_map,
+        scheduler,
+        pipeline_impl_map,
+        pipeline,
+        epilogue,
+        k_block_per_cu,
+        persistent,
+    ):
+        # Function Signature
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code = """
+
+    // Launch function
+    static float launch(const ck_tile::GemmMultiDHostArgs<DsDataType::size()>&  args, const ck_tile::stream_config& stream) {"""
+        elif self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code = """
+
+    // Launch function
+    static float launch(const ck_tile::GemmHostArgs& args, const ck_tile::stream_config& stream) {"""
+
+        # Initialization
+        instance_code += """
+        const ck_tile::index_t k_grain = args.k_batch * TileK;
+        const ck_tile::index_t K_split = (args.K + k_grain - 1) / k_grain * TileK;
+        const ck_tile::index_t num_loop = TilePartitioner::GetLoopNum(K_split);
+        const bool has_hot_loop = BaseGemmPipeline::BlockHasHotloop(num_loop);
+        const ck_tile::TailNumber tail_num = BaseGemmPipeline::GetBlockLoopTailNum(num_loop);
         
-        const auto Run = [&](const auto memory_operation_) {{
-            constexpr auto scheduler = {scheduler_type_map.get(scheduler)};
-            [[maybe_unused]] constexpr auto memory_operation = memory_operation_.value;
+        float ave_time{0};
+        """
+
+        # Runfunction body
+        instance_code += """
+        const auto Run = [&](const auto has_hot_loop_, const auto tail_number_, const auto memory_operation_) {
+            constexpr bool has_hot_loop_v = has_hot_loop_.value;
+            constexpr auto tail_number_v = tail_number_.value;"""
+
+        # Scheduler initialization
+        if self.kernel_name_prefix in ["gemm_universal", "gemm_multi_d"]:
+            instance_code += f"""
+            constexpr auto scheduler = {scheduler_type_map.get(scheduler)};"""
+        elif self.kernel_name_prefix == "gemm_preshuffle":
+            instance_code += f"""
+            constexpr auto scheduler = {scheduler_type_map.get(scheduler, "ck_tile::GemmPipelineScheduler::Default")};"""
+
+        # Memory operatio
+        instance_code += """
+            [[maybe_unused]] constexpr auto memory_operation = memory_operation_.value;"""
+
+        # UniversalGemmProblem
+        if self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code += """
 
             using UniversalGemmProblem = ck_tile::UniversalGemmPipelineProblem<
                 ADataType,
@@ -432,16 +672,154 @@ struct SelectedKernel {{
                                                 ALayout, BLayout, CLayout, TransposeC,
                                                 UseStructuredSparsity, UsePersistentKernel,
                                                 NumWaveGroups, Preshuffle>,
-                scheduler>;
-            
-            using GemmPipeline = {pipeline_impl_map.get(pipeline)}<UniversalGemmProblem>;
-            
-            // Epilogue
-"""
+                scheduler,
+                has_hot_loop_v,
+                tail_number_v>;"""
+        elif self.kernel_name_prefix == "gemm_multi_d":
+            instance_code += """
 
-        # Add epilogue configuration based on type
+            using UniversalGemmProblem = ck_tile::UniversalGemmPipelineProblem<
+                ADataType,
+                BDataType,
+                AccDataType,
+                TileShape,
+                ck_tile::TileGemmUniversalTraits<kPadM, kPadN, kPadK, DoubleSmemBuffer,
+                                                ALayout, BLayout, CLayout, TransposeC>,
+                scheduler,
+                has_hot_loop_v,
+                tail_number_v>;"""
+
+        # GemmPipeline
+        if self.kernel_name_prefix in ["gemm_universal", "gemm_multi_d"]:
+            instance_code += f"""
+
+            using GemmPipeline = {pipeline_impl_map.get(pipeline)}<UniversalGemmProblem>;"""
+        elif self.kernel_name_prefix == "gemm_preshuffle":
+            instance_code += f"""
+
+            using GemmPipeline = {pipeline_impl_map.get(pipeline, "ck_tile::WeightPreshufflePipelineAGmemBGmemCRegV2")}<UniversalGemmProblem>;"""
+
+        # Epilogue
+        instance_code += self.populate_epilogue(self.kernel_name_prefix, epilogue)
+
+        # Kernel type
+        if self.kernel_name_prefix == "gemm_multi_d":
+            instance_code += f"""
+            
+            // Kernel type
+            using GemmKernelMultiD = ck_tile::GemmKernelMultiD<TilePartitioner, GemmPipeline, GemmEpilogue>;
+            
+            // Kernel arguments
+            auto kargs = GemmKernelMultiD::MakeKernelArgs(args);
+            
+            if (!GemmKernelMultiD::IsSupportedArgument(kargs)) {{
+                throw std::runtime_error("Wrong! Arguments not supported! Skipping gemm!");
+            }}
+
+            // Get grid and block sizes
+            const dim3 grids = GemmKernelMultiD::GridSize(args.M, args.N, args.k_batch);
+            const dim3 blocks = GemmKernelMultiD::BlockSize();
+            
+            if(stream.log_level_ > 0) {{
+                std::cout << "Launching kernel with args: " << GemmKernelMultiD::GetName() << '\\n'
+                          << "grid: {{" << grids.x << ", " << grids.y << ", " << grids.z << "}}"
+                          << ", blocks: {{" << blocks.x << ", " << blocks.y << ", " << blocks.z << "}}"
+                          << std::endl;
+            }}
+            
+            // Launch kernel
+            constexpr int kBlockPerCu = {k_block_per_cu};
+            ave_time = ck_tile::launch_kernel(
+                stream,
+                ck_tile::make_kernel<kBlockPerCu>(GemmKernelMultiD{{}}, grids, blocks, 0, kargs));
+            
+            return ave_time;
+        }};"""
+
+        elif self.kernel_name_prefix in ["gemm_universal", "gemm_preshuffle"]:
+            instance_code += f"""
+            
+            // Kernel type
+            using GemmKernel = ck_tile::GemmKernel<TilePartitioner, GemmPipeline, GemmEpilogue>;
+            
+            // Kernel arguments
+            auto kargs = GemmKernel::MakeKernelArgs(args);
+            
+            if (!GemmKernel::IsSupportedArgument(kargs)) {{
+                throw std::runtime_error("Wrong! Arguments not supported! Skipping gemm!");
+            }}
+
+            // Get grid and block sizes
+            const dim3 grids = {"GemmKernel::MaxOccupancyGridSize(stream)" if persistent in [True, "true"] else "GemmKernel::GridSize(args.M, args.N, args.k_batch)"};
+            const dim3 blocks = GemmKernel::BlockSize();
+            
+            if(stream.log_level_ > 0) {{
+                std::cout << "Launching kernel with args: " << GemmKernel::GetName() << '\\n'
+                          << "grid: {{" << grids.x << ", " << grids.y << ", " << grids.z << "}}"
+                          << ", blocks: {{" << blocks.x << ", " << blocks.y << ", " << blocks.z << "}}"
+                          << std::endl;
+            }}
+            
+            // Launch kernel
+            constexpr int kBlockPerCu = {k_block_per_cu};
+            ave_time = ck_tile::launch_kernel(
+                stream,
+                ck_tile::make_kernel<kBlockPerCu>(GemmKernel{{}}, grids, blocks, 0, kargs));
+            
+            return ave_time;
+        }};"""
+
+        # Run SplitK handler
+        instance_code += """
+
+        //Run SplitK handler
+        const auto RunSplitk = [&](const auto has_hot_loop_, const auto tail_number_) {
+            if(args.k_batch == 1) {
+                Run(has_hot_loop_,
+                    tail_number_,
+                    ck_tile::integral_constant<ck_tile::memory_operation_enum,
+                                            ck_tile::memory_operation_enum::set>{});
+            } else {
+                Run(has_hot_loop_,
+                    tail_number_,
+                    ck_tile::integral_constant<ck_tile::memory_operation_enum,
+                                            ck_tile::memory_operation_enum::atomic_add>{});
+            }
+        };
+
+        BaseGemmPipeline::TailHandler(RunSplitk, has_hot_loop, tail_num);
+        return ave_time;
+    }
+};
+"""
+        return instance_code
+
+    def populate_epilogue(self, epilogue):
+        instance_code = """
+
+            // Epilogue
+            """
+
         if epilogue == "cshuffle":
-            instance_code += """            using EpilogueProblem = ck_tile::CShuffleEpilogueProblem<
+            if self.kernel_name_prefix == "gemm_universal":
+                instance_code += self.populate_cshuffle_gemm_universal()
+            elif self.kernel_name_prefix == "gemm_multi_d":
+                instance_code += self.populate_cshuffle_gemm_multi_d()
+            elif self.kernel_name_prefix == "gemm_preshuffle":
+                instance_code += self.populate_cshuffle_gemm_preshuffle()
+        else:  # default epilogue
+            if self.kernel_name_prefix == "gemm_universal":
+                instance_code += self.populate_default_gemm_universal()
+            elif self.kernel_name_prefix == "gemm_multi_d":
+                instance_code += self.populate_default_gemm_multi_d()
+            elif self.kernel_name_prefix == "gemm_preshuffle":
+                instance_code += self.populate_default_gemm_preshuffle()
+
+        return instance_code
+
+    def populate_cshuffle_gemm_universal(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::CShuffleEpilogueProblem<
                 ADataType,
                 BDataType,
                 ck_tile::tuple<>,  // DsDataType
@@ -461,10 +839,64 @@ struct SelectedKernel {{
                 memory_operation,            // MemoryOperation_
                 NumWaveGroups>;              // kNumWaveGroups_
             
-            using GemmEpilogue = ck_tile::CShuffleEpilogue<EpilogueProblem>;
-"""
-        else:  # default epilogue
-            instance_code += """            using EpilogueProblem = ck_tile::DefaultGemm2DEpilogueProblem<
+            using GemmEpilogue = ck_tile::CShuffleEpilogue<EpilogueProblem>;"""
+        return instance_code
+
+    def populate_cshuffle_gemm_multi_d(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::CShuffleEpilogueProblem<
+                ADataType,
+                BDataType,
+                DsDataType,
+                AccDataType,
+                CDataType,
+                DsLayout,
+                CLayout,
+                ElementWiseFn,
+                TilePartitioner::MPerBlock,  // kM_
+                TilePartitioner::NPerBlock,  // kN_
+                WarpPerBlock_M,              // MWave_
+                WarpPerBlock_N,              // NWave_
+                WarpTileM,                   // MPerXdl_
+                WarpTileN,                   // NPerXdl_
+                WarpTileK,                   // KPerXdl_
+                TransposeC,                  // isCTransposed_
+                memory_operation>;           // MemoryOperation_ 
+       
+            using GemmEpilogue = ck_tile::CShuffleEpilogue<EpilogueProblem>;"""
+        return instance_code
+
+    def populate_cshuffle_gemm_preshuffle(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::CShuffleEpilogueProblem<
+                ADataType,
+                BDataType,
+                ck_tile::tuple<>,  // DsDataType
+                AccDataType,
+                CDataType,
+                ck_tile::tuple<>,  // DsLayout
+                CLayout,
+                ck_tile::element_wise::PassThrough,
+                TilePartitioner::MPerBlock,  // kM_
+                TilePartitioner::NPerBlock,  // kN_
+                WarpPerBlock_M,              // MWave_
+                WarpPerBlock_N,              // NWave_
+                WarpTileM,                   // MPerXdl_
+                WarpTileN,                   // NPerXdl_
+                WarpTileK,                   // KPerXdl_
+                TransposeC,                  // isCTransposed_
+                memory_operation,            // MemoryOperation_
+                NumWaveGroups,               // kNumWaveGroups_
+                false,                       // FixedVectorSize_
+                1,                           // VectorSizeC_
+                PermuteN>;                   // isPermuteN_
+            
+            using GemmEpilogue = ck_tile::CShuffleEpilogue<EpilogueProblem>;"""
+        return instance_code
+
+    def populate_default_gemm_universal(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::DefaultGemm2DEpilogueProblem<
                 ADataType,
                 BDataType,
                 ck_tile::tuple<>,  // DsDataType
@@ -482,151 +914,60 @@ struct SelectedKernel {{
                 WarpTileK,  // kKPerXdl_
                 TransposeC>;  // isCTransposed_
             
-            using GemmEpilogue = ck_tile::DefaultGemm2DEpilogue<EpilogueProblem>;
-"""
+            using GemmEpilogue = ck_tile::DefaultGemm2DEpilogue<EpilogueProblem>;"""
+        return instance_code
 
-        instance_code += f"""
+    def populate_default_gemm_multi_d(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::DefaultGemm2DEpilogueProblem<
+                ADataType,
+                BDataType,
+                DsDataType,
+                AccDataType,
+                CDataType,
+                DsLayout,
+                CLayout,
+                ElementWiseFn,
+                TilePartitioner::MPerBlock,  // kM_
+                TilePartitioner::NPerBlock,  // kN_
+                kPadM,
+                kPadN,
+                WarpTileM,  // kMPerXdl_
+                WarpTileN,  // kNPerXdl_
+                WarpTileK,  // kKPerXdl_
+                TransposeC>;  // isCTransposed_
             
-            // Kernel type
-            using GemmKernel = ck_tile::GemmKernel<TilePartitioner, GemmPipeline, GemmEpilogue>;
+            using GemmEpilogue = ck_tile::DefaultGemm2DEpilogue<EpilogueProblem>;"""
+        return instance_code
+
+    def populate_default_gemm_preshuffle(self):
+        instance_code = """            
+            using EpilogueProblem = ck_tile::DefaultGemm2DEpilogueProblem<
+                ADataType,
+                BDataType,
+                ck_tile::tuple<>,  // DsDataType
+                AccDataType,
+                CDataType,
+                ck_tile::tuple<>,  // DsLayout
+                CLayout,
+                ck_tile::element_wise::PassThrough,
+                TilePartitioner::MPerBlock,  // kM_
+                TilePartitioner::NPerBlock,  // kN_
+                kPadM,
+                kPadN,
+                WarpTileM,  // kMPerXdl_
+                WarpTileN,  // kNPerXdl_
+                WarpTileK,  // kKPerXdl_
+                TransposeC>;  // isCTransposed_
             
-            // Make kernel arguments
-            auto kargs = GemmKernel::MakeKernelArgs(args);
-            
-            if (!GemmKernel::IsSupportedArgument(kargs)) {{
-                throw std::runtime_error("Wrong! Arguments not supported! Skipping gemm!");
-            }}
-            
-            // Get grid and block sizes
-            const dim3 grids = {"GemmKernel::MaxOccupancyGridSize(stream)" if persistent in [True, "true"] else "GemmKernel::GridSize(args.M, args.N, args.k_batch)"};
-            const dim3 blocks = GemmKernel::BlockSize();
-            
-            if(stream.log_level_ > 0) {{
-                std::cout << "Launching kernel with args: " << GemmKernel::GetName() << '\\n'
-                          << "grid: {{" << grids.x << ", " << grids.y << ", " << grids.z << "}}"
-                          << ", blocks: {{" << blocks.x << ", " << blocks.y << ", " << blocks.z << "}}"
-                          << std::endl;
-            }}
-            
-            // Launch kernel
-            constexpr int kBlockPerCu = {k_block_per_cu};
-            float ave_time = ck_tile::launch_kernel(
-                stream,
-                ck_tile::make_kernel<kBlockPerCu>(GemmKernel{{}}, grids, blocks, 0, kargs));
-            
-            return ave_time;
-        }};
-
-        float ave_time = 0.f;
-
-        if(args.k_batch == 1) {{
-            ave_time = Run(ck_tile::integral_constant<ck_tile::memory_operation_enum,
-                                        ck_tile::memory_operation_enum::set>{{}});
-        }} else {{
-            ave_time = Run(ck_tile::integral_constant<ck_tile::memory_operation_enum,
-                                        ck_tile::memory_operation_enum::atomic_add>{{}});
-        }}
-
-        return ave_time;
-    }}
-}};
-"""
-        return kernel_name, instance_code
-
-    def run(self, num_workers=None):
-        """Run the builder to generate individual kernel files"""
-        # Generate individual kernel files
-        self.generate_individual(num_workers)
-
-    def generate_individual(self, num_workers=None):
-        """Generate individual kernel files for separate compilation with parallel processing"""
-        if num_workers is None:
-            num_workers = min(
-                multiprocessing.cpu_count(), 8
-            )  # Limit to avoid memory issues
-
-        tile_configs = self._get_tile_configs()
-        trait_combos = self._generate_trait_combinations()
-        k_block_per_cu = self.config.get("k_block_per_cu")
-        if k_block_per_cu is None:
-            k_block_per_cu = 1
-
-        # Prepare work items for parallel processing
-        work_items = []
-        for tile_config in tile_configs:
-            for trait_combo in trait_combos:
-                work_items.append(
-                    (
-                        tile_config,
-                        trait_combo,
-                        k_block_per_cu,
-                        self.working_path,
-                        self.gpu_target,
-                        self.datatype,
-                        self.layout,
-                        self.config_json,
-                    )
-                )
-        print(
-            f"Generating {len(work_items)} individual kernel files using {num_workers} workers..."
-        )
-        print(f"  Tile configs: {len(tile_configs)}")
-        print(f"  Trait combinations: {len(trait_combos)}")
-        print(f"  Total kernels: {len(work_items)}")
-
-        # Show first few work items for debugging
-        if work_items:
-            print("  First work item example:")
-            tile_config, trait_combo = work_items[0][:2]
-            print(f"    Tile config: {tile_config}")
-            print(f"    Trait combo: {trait_combo[:3]}")  # Show first 3 traits
-
-        # Process work items in parallel
-        kernel_list = []
-        completed = 0
-
-        with concurrent.futures.ProcessPoolExecutor(
-            max_workers=num_workers
-        ) as executor:
-            # Submit all work items
-            print(f"  Submitting {len(work_items)} tasks to executor...")
-            future_to_item = {
-                executor.submit(_generate_single_kernel_individual, item): item
-                for item in work_items
-            }
-            print("  All tasks submitted, waiting for completion...")
-
-            # Collect results with progress reporting
-            for future in concurrent.futures.as_completed(future_to_item):
-                completed += 1
-                if completed % 100 == 0 or completed == len(work_items):
-                    print(
-                        f"  Progress: {completed}/{len(work_items)} kernels generated"
-                    )
-                try:
-                    result = future.result()
-                    if result:
-                        kernel_list.append(result)
-                except Exception as exc:
-                    item = future_to_item[future]
-                    print(f"Kernel generation failed for {item}: {exc}")
-
-        # Sort kernel list for consistent ordering
-        kernel_list.sort(key=lambda x: x[0])  # Sort by kernel name
-
-        # Generate CMake include file for individual targets
-        self._generate_cmake_individual_targets(kernel_list)
-
-        print(
-            f"Generated {len(kernel_list)} individual kernel files in {self.working_path}"
-        )
+            using GemmEpilogue = ck_tile::DefaultGemm2DEpilogue<EpilogueProblem>;"""
+        return instance_code
 
     def _generate_cmake_individual_targets(self, kernel_list):
         """Generate CMake include file that creates individual targets"""
-        cmake_code = f"""# Generated CMake file for individual GEMM targets
-# Datatype: {self.datatype}, Layout: {self.layout}
-
-"""
+        cmake_code = f"""# Generated CMake file for individual {self.kernel_name_prefix} targets
+        # Datatype: {self.datatype}, Layout: {self.layout}
+        """
 
         for kernel_name, trait_combo, tile_config in kernel_list:
             pipeline, epilogue, scheduler = trait_combo[:3]
@@ -640,187 +981,11 @@ struct SelectedKernel {{
                 str(x) for x in trait_combo[3:]
             )
 
-            cmake_code += f'create_individual_gemm_target("{self.datatype}" "{self.layout}" "{trait_str}" "{tile_str}")\n'
+            cmake_code += f'create_individual_{self.kernel_name_prefix}_target("{self.datatype}" "{self.layout}" "{trait_str}" "{tile_str}")\n'
 
         # Write CMake include file
-        with open(self.working_path / "gemm_individual_targets.cmake", "w") as f:
+        with open(
+            self.working_path / f"{self.kernel_name_prefix}_individual_targets.cmake",
+            "w",
+        ) as f:
             f.write(cmake_code)
-
-
-def _generate_single_kernel_individual(work_item):
-    """Worker function to generate a single individual kernel file"""
-    (
-        tile_config,
-        trait_combo,
-        k_block_per_cu,
-        working_path,
-        gpu_target,
-        datatype,
-        layout,
-        config_json,
-    ) = work_item
-
-    # Create a temporary builder instance for this worker
-    builder = GemmKernelBuilder(working_path, gpu_target, datatype, layout, config_json)
-
-    try:
-        kernel_name, instance_code = builder._generate_kernel_instance(
-            tile_config, trait_combo, k_block_per_cu
-        )
-
-        # Create simplified filename without the "gemm_" prefix
-        # Remove "gemm_" from the beginning of kernel_name for the filename
-        simplified_name = kernel_name
-        if simplified_name.startswith("gemm_"):
-            simplified_name = simplified_name[5:]  # Remove "gemm_" prefix
-
-        # Write individual header file
-        header_file = working_path / f"gemm_single_{simplified_name}.hpp"
-        with open(header_file, "w") as f:
-            f.write(instance_code)
-
-        return (kernel_name, trait_combo, tile_config)
-    except Exception as e:
-        print(f"Error generating individual kernel: {e}")
-        return None
-
-
-def main():
-    parser = argparse.ArgumentParser(
-        description="GEMM kernel instance builder with parallel support"
-    )
-    parser.add_argument("--working_path", required=True, help="Working directory path")
-    parser.add_argument(
-        "--gpu_target",
-        required=True,
-        help="GPU target architecture",
-    )
-    parser.add_argument(
-        "--datatype",
-        required=True,
-        choices=["fp16", "fp8", "bf16", "bf8"],
-        help="Data type",
-    )
-    parser.add_argument(
-        "--layout",
-        required=True,
-        choices=["rcr", "rrr", "ccr", "crr"],
-        help="Matrix layout",
-    )
-    parser.add_argument("--config_json", help="Configuration JSON file")
-    parser.add_argument(
-        "--num_workers", type=int, help="Number of parallel workers (default: auto)"
-    )
-    parser.add_argument(
-        "--gen_all_individual",
-        action="store_true",
-        help="Generate individual kernel files",
-    )
-    parser.add_argument(
-        "--gen_single", action="store_true", help="Generate a single kernel file"
-    )
-    parser.add_argument("--kernel_name", help="Kernel name for single generation")
-    parser.add_argument(
-        "--tile_config", help="Tile configuration string for single generation"
-    )
-    parser.add_argument(
-        "--trait_combo", help="Trait combination string for single generation"
-    )
-    parser.add_argument(
-        "--list_kernels",
-        action="store_true",
-        help="List kernel configurations without generating files",
-    )
-
-    args = parser.parse_args()
-
-    assert args.datatype in ["fp16", "bf16", "fp8", "bf8"], (
-        f"Invalid datatype string: {args.datatype} (supported datatypes are [fp16, bf16, fp8, and bf8])"
-    )
-
-    layout_parts = args.layout.lower()
-    assert len(layout_parts) == 3, (
-        f"Invalid layout string: {args.layout} (must be 3 characters like 'rcr' where r stands for row major and c stands for column major)"
-    )
-    assert layout_parts[0] in ["r", "c"] and layout_parts[1] in ["r", "c"], (
-        f"Invalid matrix_a layout : {layout_parts[0]} or matrix_b layout: {layout_parts[1]} (matrix_a and matrix_b must be either 'r' for row major or 'c' for column major)"
-    )
-    assert layout_parts[2] == "r", (
-        f"Invalid matrix_c layout: {layout_parts[2]} (must be 'r' only as currently we are supporting only row major)"
-    )
-
-    # Create builder
-    builder = GemmKernelBuilder(
-        args.working_path, args.gpu_target, args.datatype, args.layout, args.config_json
-    )
-
-    if args.list_kernels:
-        builder.write_kernel_list()
-    elif args.gen_single:
-        # Generate a single kernel file
-        if not args.kernel_name or not args.tile_config or not args.trait_combo:
-            parser.error(
-                "--gen_single requires --kernel_name, --tile_config, and --trait_combo"
-            )
-
-        # Parse tile config
-        tile_parts = args.tile_config.split("_")
-        tile_dims = tile_parts[0].split("x")
-        warp_dims = tile_parts[1].split("x")
-        warp_tile_dims = tile_parts[2].split("x")
-
-        tile_config = {
-            "tile_m": int(tile_dims[0]),
-            "tile_n": int(tile_dims[1]),
-            "tile_k": int(tile_dims[2]),
-            "warp_m": int(warp_dims[0]),
-            "warp_n": int(warp_dims[1]),
-            "warp_k": int(warp_dims[2]),
-            "warp_tile_m": int(warp_tile_dims[0]),
-            "warp_tile_n": int(warp_tile_dims[1]),
-            "warp_tile_k": int(warp_tile_dims[2]),
-        }
-
-        # Parse trait combo
-        trait_parts = args.trait_combo.split("_")
-        trait_combo = (
-            trait_parts[0],  # pipeline
-            trait_parts[1],  # epilogue
-            trait_parts[2],  # scheduler
-            trait_parts[3] == "True",  # pad_m
-            trait_parts[4] == "True",  # pad_n
-            trait_parts[5] == "True",  # pad_k
-            trait_parts[6] == "True",  # persistent
-        )
-
-        k_block_per_cu = builder.config.get("k_block_per_cu")
-        if k_block_per_cu is None:
-            k_block_per_cu = 1
-
-        # Generate the kernel
-        kernel_name, instance_code = builder._generate_kernel_instance(
-            tile_config, trait_combo, k_block_per_cu
-        )
-
-        # Write the file
-        simplified_name = kernel_name
-        if simplified_name.startswith("gemm_"):
-            simplified_name = simplified_name[5:]
-
-        header_file = builder.working_path / f"gemm_single_{simplified_name}.hpp"
-        with open(header_file, "w") as f:
-            f.write(instance_code)
-
-        print(f"Generated {header_file}")
-
-    elif args.gen_all_individual:
-        # Generate all individual kernel files
-        builder.run(args.num_workers)
-    else:
-        parser.error(
-            "Must specify one of: --list_kernels, --gen_all_individual, or --gen_single"
-        )
-
-
-if __name__ == "__main__":
-    main()
