+++ tile_engine/ops/gemm/gemm_multi_d/configs/default_config.json
