+++ tile_engine/ops/gemm/gemm_multi_d/gemm_multi_d_benchmark_single.cpp
@@ -11,12 +11,12 @@
 
 #include "ck_tile/core.hpp"
 #include "ck_tile/host.hpp"
-#include "ck_tile/ops/common/utils.hpp"
-#include "gemm_preshuffle_profiler.hpp"
-#include "gemm_preshuffle_common.hpp"
+#include "gemm_multi_d_profiler.hpp"
+#include "gemm_multi_d_common.hpp"
 
 // The kernel header is included via the compile command line with -include flag
 // It defines SelectedKernel struct and KERNEL_NAME
+// DataTypeTraits are now defined in gemm_multi_d_common.hpp
 
 // Create argument parser
 inline auto create_args(int argc, char* argv[])
@@ -27,12 +27,13 @@ inline auto create_args(int argc, char* argv[])
         .insert("k", "2048", "The value for k dimension. Default is 2048.")
         .insert("stride_a", "0", "The stride value for tensor A. Default is 0.")
         .insert("stride_b", "0", "The stride value for tensor B. Default is 0.")
+        .insert("stride_ds", "0", "The stride value for tensor Ds . Default is 0.")
         .insert("stride_c", "0", "The stride value for tensor C. Default is 0.")
         .insert("split_k", "1", "The split value for k dimension. Default is 1.")
         .insert("verify",
-                "2",
-                "The type of validation. Set to 0 for no validation, 1 for validation on CPU, or 2 "
-                "for validation on GPU. Default is 0, no validation.")
+                "1",
+                "for validation on GPU. Default is 1, validation on CPU, as validation on GPU is "
+                "not supported.")
         .insert("log",
                 "false",
                 "Whether output kernel instance information or not. Possible values are true or "
@@ -77,34 +78,43 @@ inline auto create_args(int argc, char* argv[])
 
 void benchmark_single(const ck_tile::ArgParser& arg_parser)
 {
-    // Use ck_tile::DataTypeTraits to get the actual type names from the generated header
+    // Use DataTypeTraits to get the actual type names from the generated header
     // The generated header defines ADataType, BDataType, AccDataType, CDataType
-    std::string dtype_a   = ck_tile::DataTypeTraits<ADataType>::name;
-    std::string dtype_b   = ck_tile::DataTypeTraits<BDataType>::name;
-    std::string dtype_acc = ck_tile::DataTypeTraits<AccDataType>::name;
-    std::string dtype_c   = ck_tile::DataTypeTraits<CDataType>::name;
+    std::string dtype_a   = DataTypeTraits<ADataType>::name;
+    std::string dtype_b   = DataTypeTraits<BDataType>::name;
+    std::string dtype_acc = DataTypeTraits<AccDataType>::name;
+    std::string dtype_c   = DataTypeTraits<CDataType>::name;
+    std::string dtype_d0  = DataTypeTraits<D0DataType>::name;
+    std::string dtype_d1  = DataTypeTraits<D1DataType>::name;
 
     // Layout names from the layout types
-    std::string layout_a = ALayout::name;
-    std::string layout_b = BLayout::name;
-    std::string layout_c = CLayout::name;
+    std::string layout_a  = ALayout::name;
+    std::string layout_b  = BLayout::name;
+    std::string layout_c  = CLayout::name;
+    std::string layout_d0 = D0Layout::name;
+    std::string layout_d1 = D1Layout::name;
 
-    // Create GemmProblem struct
-    GemmProblem gemm_problem{arg_parser.get_int("split_k"),
-                             arg_parser.get_int("m"),
-                             arg_parser.get_int("n"),
-                             arg_parser.get_int("k"),
-                             arg_parser.get_int("stride_a"),
-                             arg_parser.get_int("stride_b"),
-                             arg_parser.get_int("stride_c"),
-                             dtype_a,
-                             dtype_b,
-                             dtype_acc,
-                             dtype_c,
-                             layout_a,
-                             layout_b,
-                             layout_c,
-                             arg_parser.get_bool("structured_sparsity")};
+    // Create GemmMultiDProblem struct
+    GemmMultiDProblem gemm_multi_d_problem{arg_parser.get_int("split_k"),
+                                           arg_parser.get_int("m"),
+                                           arg_parser.get_int("n"),
+                                           arg_parser.get_int("k"),
+                                           arg_parser.get_int("stride_a"),
+                                           arg_parser.get_int("stride_b"),
+                                           arg_parser.get_int("stride_ds"),
+                                           arg_parser.get_int("stride_ds"),
+                                           arg_parser.get_int("stride_c"),
+                                           dtype_a,
+                                           dtype_b,
+                                           dtype_d0,
+                                           dtype_d1,
+                                           dtype_acc,
+                                           dtype_c,
+                                           layout_a,
+                                           layout_b,
+                                           layout_d0,
+                                           layout_d1,
+                                           layout_c};
 
     // Create Setting struct
     Setting setting{arg_parser.get_int("warmup"),
@@ -119,29 +129,18 @@ void benchmark_single(const ck_tile::ArgParser& arg_parser)
                     arg_parser.get_bool("json_output")};
 
     // Get the profiler instance
-    auto& profiler = GemmProfiler::instance(setting);
+    auto& profiler = GemmMultiDProfiler::instance(setting);
 
     try
     {
         // Create a lambda that wraps the kernel launch
-        std::tuple<int, int, int> warp_tile_dims = std::make_tuple(
-            SelectedKernel::WarpTileM, SelectedKernel::WarpTileN, SelectedKernel::WarpTileK);
-        std::tuple<int, int, int> tile_dims =
-            std::make_tuple(SelectedKernel::TileM, SelectedKernel::TileN, SelectedKernel::TileK);
-        std::tuple<int, int, int> warp_dims = std::make_tuple(SelectedKernel::WarpPerBlock_M,
-                                                              SelectedKernel::WarpPerBlock_N,
-                                                              SelectedKernel::WarpPerBlock_K);
-        bool permuteN                       = SelectedKernel::PermuteN;
-
-        KernelConfig config{tile_dims, warp_dims, warp_tile_dims, permuteN};
-
-        auto kernel_func = [](const ck_tile::GemmHostArgs& args,
+        auto kernel_func = [](const ck_tile::GemmMultiDHostArgs<DsDataType::size()>& args,
                               const ck_tile::stream_config& stream) {
             return SelectedKernel::launch(args, stream);
         };
 
         // Benchmark the kernel
-        profiler.benchmark(gemm_problem, kernel_func, config);
+        profiler.benchmark(gemm_multi_d_problem, kernel_func);
 
         // Select best instance based on metric
         profiler.select_best_instance(static_cast<Metric>(arg_parser.get_int("metric")));
