+++ tile_engine/ops/gemm/gemm_validation_utils.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
-# Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
 # SPDX-License-Identifier: MIT
+# Copyright (c) 2024, Advanced Micro Devices, Inc. All rights reserved.
 
 import logging
 from typing import Tuple, List
@@ -660,7 +660,6 @@ def validate_whole_wg_cover_configuration(
         )
 
         if not wg_cover_core_valid:
-            print("I am here 3")
             logging.debug(
                 f"whole workgroup cover failed for Matrix B distribution: {wg_cover_core_error}"
             )
